refactor: slightly refactor everything
