--- discord/api/execution/__init__.py
@@ -1,6 +1,0 @@
-"""
-discord.execution
-~~~~~~~~~~~~~~~~~
-Rate Limit execution for the Discord API
-"""
-from .executer import *
