--- discord/banner.txt
@@ -1,11 +1,0 @@
-
-
-      ${bold_red}888 d8b                                       888     d8b              ${reset}${bold_white}git sha1:  ${reset}${thin_white}${git_sha}${reset}
-      ${red}888 Y8P                                       888     Y8P              ${reset}${bold_white}version:   ${reset}${thin_white}${version}${reset}
-      ${yellow}888                                           888                      ${reset}${bold_white}copyright: ${reset}${thin_white}${copyright} | ${license} ${reset}
-  ${green}.d88888 888 .d8888b   .d8888b .d88b.  888d888 .d88888     888  .d88b.      ${reset}${bold_white}date:      ${reset}${thin_white}${current_time} ${reset}
- ${blue}d88' 888 888 88K      d88P'   d88''88b 888P'  d88' 888     888 d88''88b     ${reset}${bold_white}python:    ${reset}${thin_white}${py_version} ${reset}
- ${cyan}888  888 888 Y88888b. 888     888  888 888    888  888     888 888  888     ${reset}${bold_white}api:       ${reset}${thin_white}v10 ${reset}
- ${purple}Y88b 888 888      X88 Y88b.   Y88..88P 888    Y88b 888 d8b 888 Y88..88P     ${reset}${bold_white}discord:   ${reset}${thin_white}https://discord.gg/cvCAwntVhm ${reset}
-  ${bold_purple}Y888888 888  88888P'  'Y8888P 'Y88P'  888     'Y88888 Y8P 888  'Y88P'      ${reset}${bold_white}docs:      ${reset}${thin_white}https://discord.rtfd.io/en/latest ${reset}
-
