+++ discord/panels/banner.txt
@@ -1,0 +1,12 @@
+
+
+             ${bold_red})8888  88888                                                                                    8888
+             ${red})8888   PPPP                                                                                    8888                ${bold_red}..${reset}            ${purple}o888888o..${reset}
+   ${yellow}..o88888..)8888  .oooo    .o8888888P..       .o8888888o.       ..o8888888o.     .oooo .o8888    .o88888o. 8888             ${bold_red}.8888K${reset}           ${purple}888888888888.${reset}
+ ${green}.8888888888888888  88888  .88888PPPK88888.  .88888888888888K   .888888888888888   888888888888  8888888888888888             ${bold_red}88888K${reset}   ${purple}.88888"         "88888.${reset}
+${blue}.8888"       88888  88888  88888..    PPPPP  8888P       88888 .8888P       88888  888888       88888       88888             ${bold_red}88888K${reset}   ${purple}88888             88888${reset}
+${cyan}8888P         8888  88888   888888888888o.  (8888              8888P         8888P 88888       )8888         8888             ${bold_red}88888K${reset}   ${purple}88888             (8888${reset}
+${yellow}88888        (8888  88888       ""PP8888888 )8888        .oooo 88888        .8888P 88888        8888         8888             ${bold_red}88888K${reset}    ${purple}88888.          o88888${reset}
+ ${green}88888..  ..888888  88888  8888K      88888  88888..   .88888"  88888..   ..8888P  88888        88888o.   ..88888    ${bold_purple}88888${reset}    ${bold_red}88888K${reset}     ${purple}8888888.....o888888${reset}
+  ${bold_red}Y888888888888888  88888   88888888888888     888888888888P     "888888888888P    88888          888888888888888    ${bold_purple}88888${reset}    ${bold_red}88888K${reset}        ${purple}Y88888888888P${reset}
+      ${red}""""                      """"""             """""              """""                          """""
