+++ extras/docs.txt
@@ -1,0 +1,1 @@
+sphinx~=4.4.0
