+++ extras/speed.txt
@@ -1,0 +1,4 @@
+aiodns~=3.0
+msgspec~=0.11.0
+Brotli~=1.0.9
+cchardet~=2.1.7
